feat(widget): refacotored set operation hook
